Refactor -- Table 검색테스트 테스트 모델 추가, tables테스트로 이동.
